+++ inst/shinyApp/ui.R
@@ -1,11 +1,10 @@
 shiny::shinyUI(shiny::fluidPage(
-  shiny::titlePanel('mRchmadness: 2017'),
+  shiny::titlePanel('mRchmadness: 2018'),
 
   shiny::sidebarLayout(
     shiny::sidebarPanel(
       shiny::selectInput('sex', 'Men or Women:',
-                         choices = c('Men',
-                                     'Women')),
+                         choices = c('Men')),
 
       shiny::selectInput('scoring', 'Scoring Scheme:',
                          choices = c('Traditional',
@@ -52,6 +51,7 @@ shiny::shinyUI(shiny::fluidPage(
         shiny::a('here', href = 'https://saberpowers.shinyapps.io/mRchmadness/'),
         'via shinyapps.io, you can run the Shiny app locally if you want more computational muscle.'
       ),
+      shiny::p('NOTE: The bracket is men\'s only for the moment, pending the release of women\'s population picks from ESPN. We will add support for the women\'s bracket as soon as those data are available. This is as of the evening of March 12th, 2018.', style='color: red;'),
       shiny::br(),
       shiny::tabsetPanel(
         shiny::tabPanel('Empty Bracket',